Lots of changes
Made some small changes to the main classes. Fixed updating of yPosition so that drawn dominos appear one line below once there is no more horizontal space on the window. Added 2 classes. CanvasTest and DrawDominoV2, which were used for trying to create my own canvas and using it in the domino drawing process. Notes can be found in source code.